requirements.txt
backcall==0.2.0
bcrypt==4.2.0
click==8.1.7
colorama==0.4.6
decorator==5.1.1
Flask==2.2.5
Flask-Bcrypt==1.0.1
Flask-SQLAlchemy==3.0.5
Flask-WTF==1.1.1
greenlet==3.1.1
importlib-metadata==6.7.0
ipython==7.34.0
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.4
MarkupSafe==2.1.5
matplotlib-inline==0.1.6
parso==0.8.4
pickleshare==0.7.5
prompt_toolkit==3.0.48
psycopg2-binary==2.9.9 
Pygments==2.17.2
python-dotenv==0.21.1
SQLAlchemy==2.0.36
traitlets==5.9.0
typing_extensions==4.7.1
wcwidth==0.2.13
Werkzeug==2.2.3
WTForms==3.0.1
zipp==3.15.0
